UXD-805 docs(ListBox): restructure single and multi listbox
