+++ packages/ListBox/stories/ListBox.examples.stories.js
@@ -3,13 +3,35 @@ import { getStoryName } from "storybook/storyTree";
 import ExampleStory from "storybook/components/ExampleStory";
 import OnChange from "./examples/OnChange";
 import Lazy from "./examples/LazyListBox/Lazy";
+import Subcomponents from "./examples/Subcomponents";
 import ListBox from "../src";
 
 const storyName = getStoryName("ListBox");
 
+const paramaters = {
+  viewMode: "story",
+  options: {
+    isToolshown: false,
+    showPanel: true,
+  },
+};
+
 export default {
   title: `${storyName}/Examples`,
   component: ListBox,
+};
+
+export const SubcomponentsStory = () => (
+  <ExampleStory component="ListBox" storyName="Subcomponents" fileName="examples/Subcomponents.js">
+    <Subcomponents />
+  </ExampleStory>
+);
+
+SubcomponentsStory.story = {
+  name: "Subcomponents",
+  parameters: {
+    ...paramaters,
+  },
 };
 
 export const OnChangeStory = () => (
@@ -18,8 +40,22 @@ export const OnChangeStory = () => (
   </ExampleStory>
 );
 
+OnChangeStory.story = {
+  name: "OnChange",
+  parameters: {
+    ...paramaters,
+  },
+};
+
 export const LazyStory = () => (
   <ExampleStory component="ListBox" storyName="Lazy ListBox" fileName="examples/Lazy.js">
     <Lazy />
   </ExampleStory>
 );
+
+LazyStory.story = {
+  name: "Lazy Load API",
+  parameters: {
+    ...paramaters,
+  },
+};
